Backend/api/requirements.txt
fastapi
uvicorn
pydantic
httpx
pymupdf
multipart
python-multipart
python-dotenv